fix: extract commom
