+++ cmd/main.go
@@ -17,7 +17,7 @@ func main() {
 	err := godotenv.Load(".env")
 
 	if err != nil {
-		log.Fatalf("Error loading .env file")
+		log.Fatalf("Error loading .env file: %s", err)
 	}
 
 	addr := os.Getenv("PORT")
