+++ commom/hash.go
